+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Features:
   * Internal elements are now marked with `@internal` annotation in Dart generated code.
+  * Values assigned to constants or fields are now validated for type correctness.
 
 ## 10.0.0
 Release date: 2021-10-01